requirements.txt
torch
torchvision
transformers
chromadb
Pillow
sentencepiece
fastapi
uvicorn
python-multipart
jinja2
sentence-transformers
PyMuPDF
python-docx
python-pptx
watchdog
pystray
PyQt5
